Pull out common code from stream last for sorted first/last.
